+++ BlueprintUI/Tests/ElementContentTests.swift
@@ -3,6 +3,22 @@ import XCTest
 
 class ElementContentTests: XCTestCase {
 
+    func test_measurement_caching() {
+        
+        let environment = Environment.empty
+        
+        let element = MeasurableElement()
+        
+        _ = element.content.measure(in: .unconstrained, environment: environment)
+        XCTAssertEqual(MeasurableElement.measureCount, 1)
+        
+        _ = element.content.measure(in: .unconstrained, environment: environment)
+        XCTAssertEqual(MeasurableElement.measureCount, 1)
+        
+        _ = element.content.measure(in: .unconstrained, environment: environment)
+        XCTAssertEqual(MeasurableElement.measureCount, 1)
+    }
+    
     func test_noChildren() {
         let container = ElementContent(layout: FrameLayout())
         XCTAssertEqual(container.childCount, 0)
@@ -55,6 +71,22 @@ class ElementContentTests: XCTestCase {
             CGSize(width: 600, height: 800))
     }
     
+}
+
+fileprivate struct MeasurableElement : Element {
+    
+    static var measureCount : Int = 0
+    
+    var content: ElementContent {
+        ElementContent(measurementCachingKey: .init(type: Self.self, input: "element")) { constraint -> CGSize in
+            Self.measureCount += 1
+            return .init(width: 20.0, height: 20.0)
+        }
+    }
+    
+    func backingViewDescription(bounds: CGRect, subtreeExtent: CGRect?) -> ViewDescription? {
+        UIView.describe { _ in }
+    }
 }
 
 fileprivate struct SimpleElement: Element {